+++ output/strapi5-schemas/app/controllers/app.js
@@ -1,0 +1,2 @@
+import { createCoreController } from '@strapi/strapi/factories';
+export default createCoreController('api::app.app');