+++ output/strapi5-schemas/article/routes/article.js
@@ -1,0 +1,2 @@
+import { createCoreRouter } from '@strapi/strapi/factories';
+export default createCoreRouter('api::article.article');