+++ output/strapi5-schemas/article/services/article.js
@@ -1,0 +1,2 @@
+import { createCoreService } from '@strapi/strapi/factories';
+export default createCoreService('api::article.article');