+++ output/strapi5-schemas/dataset/controllers/dataset.js
@@ -1,0 +1,2 @@
+import { createCoreController } from '@strapi/strapi/factories';
+export default createCoreController('api::dataset.dataset');